Feature: background colour on request option (#241)
* creates a badge surrounding the method

* abstracts to a reusable function

* interchanges the GET and POST request colours

* changes method name, removes redundant .tostring

* realigns the text in the badge
